+++ controllers/api/categories.js
@@ -12,15 +12,17 @@ async function index(req, res) {
 }
 
 async function createCat(req, res) {
-  
-  console.log(req.body);
   let catData = req.body;
+  catData.user = req.user;
+  console.log(catData.user)
   let newCat = await Category.create(catData);
   return res.json(newCat)
 }
 
 async function deleteCat(req, res) {
-  const removedCat = await Category.findByIdAndRemove(req.params.id);
-  console.log(req.params.id)
-  res.status(200).json(removedCat);
+  const owner = await Category.findById(req.params.id);
+  if (owner.user.toString() === req.user._id) {
+    const removedCat = await Category.findByIdAndRemove(req.params.id);
+    res.status(200).json(removedCat);
+  } 
 }
