+++ controllers/api/decks.js
@@ -14,15 +14,18 @@ async function index(req, res) {
 
 async function createDeck(req, res) {
   let deckData = req.body;
+  deckData.user = req.user;
   let newDeck = await Deck.create(deckData);
   return res.json(newDeck)
 }
 
-//obsolete but when deleted, deckspage crashes.
+//obsolete but when deleted, DecksPage crashes.
 async function addCard(req, res) {}
 
 async function deleteDeck(req, res) {
-  const removedDeck = await Deck.findByIdAndRemove(req.params.id);
-  console.log(req.params.id)
-  res.status(200).json(removedDeck);
+  const owner = await Deck.findById(req.params.id);
+  if (owner.user.toString() === req.user._id) {
+    const removedDeck = await Deck.findByIdAndRemove(req.params.id);
+    res.status(200).json(removedDeck);
+  }
 }
